# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAS API",
    "version": "0.4.0",
    "description": "Pinterest → Telegram Traffic Acquisition System. Error envelope: {error:{code,message,details?}}. Коды: VALIDATION_ERROR 400, UNAUTHORIZED 401, FORBIDDEN 403, NOT_FOUND 404, CONFLICT 409, UNPROCESSABLE 422, RATE_LIMITED 429, INTERNAL 500."
  },
  "servers": [{ "url": "{PUBLIC_BASE_URL}/api/v1", "variables": { "PUBLIC_BASE_URL": { "default": "https://tvrs.io" } } }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "UNPROCESSABLE", "RATE_LIMITED", "INTERNAL"] },
              "message": { "type": "string" },
              "details": {}
            }
          }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness/readiness",
        "responses": { "200": { "description": "ok" }, "503": { "description": "degraded" } }
      }
    },
    "/events": {
      "post": {
        "summary": "Beacon-событие моста (только telegram_click; link_click/bridge_view пишутся сервером)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "properties"],
                "properties": {
                  "name": { "type": "string", "enum": ["telegram_click"] },
                  "properties": {
                    "type": "object",
                    "required": ["slug", "token"],
                    "properties": {
                      "slug": { "type": "string", "maxLength": 96 },
                      "token": { "type": "string", "maxLength": 64 },
                      "session_id": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "accepted (best-effort)" },
          "400": { "description": "VALIDATION_ERROR", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "403": { "description": "server-side only event", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "429": { "description": "rate limit (60/min/IP)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/admin/auth/login": {
      "post": {
        "summary": "Логин админа (argon2id + TOTP; lockout 5×15мин; rate limit 20/15мин/IP)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "totp"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "totp": { "type": "string", "pattern": "^\\d{6}$" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "token": { "type": "string" }, "expiresIn": { "type": "integer", "example": 900 } } }
              }
            }
          },
          "401": { "description": "invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
          "429": { "description": "locked / too many", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
        }
      }
    },
    "/admin/me": {
      "get": {
        "summary": "Whoami",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "principal" }, "401": { "description": "unauthorized" } }
      }
    },
    "/admin/tracking-links": {
      "post": {
        "summary": "Издать tracking-ссылку (editor+). Поддерживает Idempotency-Key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "schema": { "type": "string", "maxLength": 128 } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["source_id", "landing_slug"],
                "properties": {
                  "source_id": { "type": "string" },
                  "landing_slug": { "type": "string" },
                  "pin_id": { "type": "string", "pattern": "^\\d+$" },
                  "campaign_id": { "type": "string", "pattern": "^\\d+$" },
                  "cluster_id": { "type": "string", "pattern": "^\\d+$" },
                  "keyword_id": { "type": "string", "pattern": "^\\d+$" },
                  "creative_variant": { "type": "string" },
                  "landing_variant": { "type": "string" },
                  "placement": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created (или replay с Idempotency-Replayed: true)",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "short_code": { "type": "string" }, "url": { "type": "string" }, "tg_url": { "type": "string" } } }
              }
            }
          },
          "401": { "description": "unauthorized" },
          "403": { "description": "viewer cannot create" }
        }
      },
      "get": {
        "summary": "Последние tracking-ссылки со счётчиками (viewer+)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }],
        "responses": { "200": { "description": "list" }, "401": { "description": "unauthorized" } }
      }
    },
    "/admin/pins": {
      "get": {
        "summary": "Контент-очередь (viewer+)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["idea", "approved", "scheduled", "published", "paused"] } },
          { "name": "cluster_id", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "list" } }
      },
      "post": {
        "summary": "Создать контент-единицу (status=idea; editor+)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cluster_id", "title"],
                "properties": {
                  "cluster_id": { "type": "string", "pattern": "^\\d+$" },
                  "keyword_id": { "type": "string" },
                  "campaign_id": { "type": "string" },
                  "title": { "type": "string", "minLength": 3 },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "created" }, "403": { "description": "viewer cannot create" } }
      }
    },
    "/admin/pins/{id}": {
      "patch": {
        "summary": "Сменить статус (переходы: idea→approved→scheduled→published; published↔paused; editor+)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["status"], "properties": { "status": { "type": "string", "enum": ["idea", "approved", "scheduled", "published", "paused"] } } } } }
        },
        "responses": { "200": { "description": "updated" }, "422": { "description": "invalid transition" } }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "Список пользователей (viewer+)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "state", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "list + total" } }
      }
    },
    "/admin/users/{id}": {
      "get": {
        "summary": "Карточка пользователя (атрибуция, сегменты, события)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "card" }, "404": { "description": "not found" } }
      }
    }
  }
}